refactor(queue): extract gate-check policy/publish/audit functions into their own module
Part of #4013's module-split sequence (step 8, after transient-locks.ts,
signal-snapshot.ts, duplicate-detection.ts, slop-detection.ts,
review-evasion.ts, ci-resolution.ts, and retention.ts): gateCheckPolicy,
recordPublishedGateCheckSummary, and auditGateCheckPermissionMissing
move to src/queue/gate-checks.ts. Pure mechanical move, no behavior
change. These three were never physically adjacent in the original
file (interspersed with unrelated linked-issue/pre-merge-check/public-
surface helpers that stay behind), but share no state and have no
caller besides processors.ts's own many disposition/publish call
sites, so they group cleanly by concern here.

A re-export shim keeps this file's own internal callers and the
existing test/unit/gate-check-policy.test.ts /
test/unit/repository-settings-enforcement.test.ts imports working
unchanged.
